requirements.txt
flask
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
groq
pypdf
